Fixes in Firebase Notifications and Performance Optimizing in RecyclerView: Now, when responding or liking the same confession, a new notification isn’t triggered; instead, the existing notification is updated. The same applies to liking responses to the same confession; again, the existing notification is updated (utilizing the ID of the existing notification, only altering the timestamp to ensure the notification appears at the top without displaying it repeatedly).
Similarly, in follows, if there exists a notification in the notifications collection with the same userID, fromUserID, and description ("followed you"), that particular notification is updated. This prevents multiple instances of the same follow notification from appearing in the notifications fragment when a user follows and unfollows repeatedly.

Additionally, if a user unfollows another, the 'followed you' notification related to that action is deleted from the other user's notifications. This behavior also extends to actions such as deleting confessions, retracting confession likes, removing answers, and retracting answer likes, checked against the 'confessions' and 'followings' collections.

When fetching notifications in the notifications fragment, we first check these collections to confirm whether the confessions still exist, if likes are still present, and if, for example, 'x' in the 'x followed y' notification is still among the followers. These methods ensure synchronization and accuracy between notifications and actions within the app, preventing users from being inundated with incorrect notifications.

In this commit, 'android:hardwareAccelerated="true"' was added to the manifest for performance enhancement purposes, and a placeholder was added to Glide in the confessionListAdapter.